Move renderer to a dedicated class
