+++ src/main/kotlin/fr/o80/release/render/Renderer.kt
@@ -1,0 +1,7 @@
+package fr.o80.release.render
+
+import fr.o80.release.Change
+
+interface Renderer {
+    fun render(versionName: String, changesByType: Map<String, List<Change>>): String
+}
